Put icon and bg color on Chrome's splash screen
When the game is added to the Home Screen, Chrome shows
a splash screen while the page is loading. To put an icon
on that splash screen, the manifest needs to contain at
least icons in 192x192 and 512x512:
https://developers.google.com/web/fundamentals/web-app-manifest/

Also adds a background and theme color for the splash
screen.